move bridges table to common components
